+++ src/core/controls.cpp
@@ -37,10 +37,10 @@
 #if (ENC_DT!=255 && ENC_CLK!=255) || (ENC2_DT!=255 && ENC2_CLK!=255)
   #include <AiEsp32RotaryEncoder.h>
   #if (ENC_DT!=255 && ENC_CLK!=255)
-    AiEsp32RotaryEncoder encoder = AiEsp32RotaryEncoder(ENC_DT, ENC_CLK, ENC_STEPS, ENC_PULLUP);
+    AiEsp32RotaryEncoder encoder = AiEsp32RotaryEncoder(ENC_DT, ENC_CLK, ENC_SW, -1, ENC_STEPS, !ENC_PULLUP);
   #endif
   #if (ENC2_DT!=255 && ENC2_CLK!=255)
-    AiEsp32RotaryEncoder encoder2 = AiEsp32RotaryEncoder(ENC2_DT, ENC2_CLK, ENC_STEPS, ENC2_PULLUP);
+    AiEsp32RotaryEncoder encoder2 = AiEsp32RotaryEncoder(ENC2_DT, ENC2_CLK, ENC2_SW, -1, ENC2_STEPS, !ENC2_PULLUP);
   #endif
 #endif