+++ src/core/options.h
@@ -2,7 +2,7 @@
 #define options_h
 #pragma once
 
-#define RADIOVERSION "2026.06.15"
+#define RADIOVERSION "2026.06.18"
 
 /*******************************************************
 THIS FILE IS THE DEFINITIVE HANDLER OF COMPILE OPTIONS.
@@ -78,6 +78,8 @@ https://trip5.github.io/ehRadio/myoptions/generator.html
 #endif
 #if DSP_MODEL==DSP_DUMMY
   #define DUMMYDISPLAY
+  #define DSP_WIDTH  0
+  #define DSP_HEIGHT 0
 #endif
 
 /* --- DISPLAY RESOLUTION --- */
@@ -201,7 +203,7 @@ https://trip5.github.io/ehRadio/myoptions/generator.html
 #endif
 #ifndef AUTOBACKLIGHT // Auto-dimming with the light sensor (not related to DSP_DIMMING_ENABLED)
   #ifndef AUTOBACKLIGHT_MAX
-    #define AUTOBACKLIGHT_MAX 2500
+    #define AUTOBACKLIGHT_MAX 1500
   #endif
   #ifndef AUTOBACKLIGHT_MIN
     #define AUTOBACKLIGHT_MIN 12
@@ -236,10 +238,6 @@ https://trip5.github.io/ehRadio/myoptions/generator.html
       #define BIG_BOOT_LOGO true
     #endif
   #endif
-#endif
-
-#ifndef USE_FBUFFER
-  #define USE_FBUFFER true // framebuffer: best to leave this on (will use PSRAM if available and SRAM if not)
 #endif
 
 /* Define your clock/volume page font as #define CLOCKFONT CHUNKY6 */
@@ -376,7 +374,7 @@ https://trip5.github.io/ehRadio/myoptions/generator.html
 #ifndef VS1053_RST
   #define VS1053_RST -1 // set to -1 if connected to Esp EN pin
 #endif
-// myoptions.h should also define which SPI bus will be used:
+// myoptions.h should also define which SPI bus will be used - VS1053 should be alone on this bus
 // #define VS1053_SPI 'A' // VS1053 decoder on Bus A (no display on A)
 // #define VS1053_SPI 'B' // VS1053 decoder on Bus B
 #if VS1053_CS!=255 && !defined(VS1053_SPI)
@@ -475,6 +473,15 @@ https://trip5.github.io/ehRadio/myoptions/generator.html
   #define PLAYER_FORCE_MONO false // mono option - false stereo, true mono (also add for mono I2S decoders for the VU)
 #endif
 
+/* --- VOLUME SCALE --- */
+#ifndef VOLUME_SCALE // Default 42 doubles future upstream library's scale of 21; 50-100 gives extremely fine control
+  #define VOLUME_SCALE 42
+#endif
+#if (VOLUME_SCALE < 21) || (VOLUME_SCALE > 255)
+  #error VOLUME_SCALE must be between 21 and 255
+#endif
+
+
 /* --- CHECK DECODER OPTIONS --- */
 #if (I2S_DOUT!=255) && (VS1053_CS!=255)
   #error define error in myoptions.h: both I2S_DOUT and VS1053_CS are active - set I2S_DOUT 255 or VS1053_CS 255 to disable one
@@ -563,11 +570,11 @@ https://trip5.github.io/ehRadio/myoptions/generator.html
   #define ENC_SW_PULLUP true
 #endif
 #ifdef ENC_STEPS
-  #if !(ENC_STEPS == 1  ENC_STEPS == 2 | ENC_STEPS == 4 // 1 is acceptable but 2 for more accuracy
-    #error ENC_STEPS 1 or 2 or 4 only please
+  #if !(ENC_STEPS == 1 || ENC_STEPS == 2 || ENC_STEPS == 4) // 1 is acceptable but 2 for more accuracy
+    #error ENC_STEPS 1 or 2 or 4 only please (2 is default)
   #endif
 #else
-  #define ENC_STEPS 1
+  #define ENC_STEPS 2
 #endif
 
 #ifndef ENC2_DT
@@ -587,10 +594,10 @@ https://trip5.github.io/ehRadio/myoptions/generator.html
 #endif
 #ifdef ENC2_STEPS
   #if !(ENC2_STEPS==1  ENC2_STEPS==2 | ENC2_STEPS==4 // 1 is acceptable but 2 for more accuracy
-    #error ENC2_STEPS 1 or 2 or 4 only please
+    #error ENC2_STEPS 1 or 2 or 4 only please (2 is default)
   #endif
 #else
-  #define ENC2_STEPS 1
+  #define ENC2_STEPS 2
 #endif
 
 /* --- BUTTONS --- */
@@ -607,8 +614,8 @@ https://trip5.github.io/ehRadio/myoptions/generator.html
 #ifndef BTN_DOWN
   #define BTN_DOWN 255
 #endif
-#ifndef BTN_PREV_PULLUP
-  #define BTN_PREV_PULLUP true
+#ifndef BTN_DOWN_PULLUP
+  #define BTN_DOWN_PULLUP true
 #endif
 #ifndef BTN_PLAY
   #define BTN_PLAY 255
@@ -619,20 +626,20 @@ https://trip5.github.io/ehRadio/myoptions/generator.html
 #ifndef BTN_UP
   #define BTN_UP 255
 #endif
-#ifndef BTN_NEXT_PULLUP
-  #define BTN_NEXT_PULLUP true
+#ifndef BTN_UP_PULLUP
+  #define BTN_UP_PULLUP true
 #endif
 #ifndef BTN_PREV
   #define BTN_PREV 255
 #endif
-#ifndef BTN_UP_PULLUP
-  #define BTN_UP_PULLUP true
+#ifndef BTN_PREV_PULLUP
+  #define BTN_PREV_PULLUP true
 #endif
 #ifndef BTN_NEXT
   #define BTN_NEXT 255
 #endif
-#ifndef BTN_DOWN_PULLUP
-  #define BTN_DOWN_PULLUP true
+#ifndef BTN_NEXT_PULLUP
+  #define BTN_NEXT_PULLUP true
 #endif
 #ifndef BTN_MODE
   #define BTN_MODE 255
@@ -770,9 +777,15 @@ https://trip5.github.io/ehRadio/myoptions/generator.html
 #if WIFI_ATTEMPTS < 1
   #error define error in myoptions.h: WIFI_ATTEMPTS must be at least 1
 #endif
+// How many seconds to wait after boot completed to Start the Async Services (other services are starting and too soon can crash netserver stack).  The less-important services started here (can take a full minute to complete if server connection issues):
+// verifies/downloads locale JSON file, checks for new version (triggers autoupdate), downloads PLAYLIST_DEFAULT_URL (if set), updates timezones.json.gz and rb_srvrs.json, cleans up stale search results
+#ifndef STARTUP_ASYNC_SERVICES_DELAY
+  #define STARTUP_ASYNC_SERVICES_DELAY 20
+#endif
  // How many seconds to wait after boot completed (including smart start) to consider successful, if rebooted during that time, will enter safe mode (disables smart start, autoupdate)
-#ifndef BOOT_SAFE_TIME
-  #define  BOOT_SAFE_TIME 10
+ // Make it about a minute?  It may mark boot as stable even if the Startup Services haven't finished yet
+#ifndef BOOT_STABLE_TIME
+  #define BOOT_STABLE_TIME 30
 #endif
 #ifndef SEARCHRESULTS_BUFFER
  // Buffer for chunked HTTP transfers from radio-browser.info. Defined in KB; conversion to bytes done in netserver.cpp.
@@ -789,7 +802,7 @@ https://trip5.github.io/ehRadio/myoptions/generator.html
   #define SEARCHRESULTS_YIELDINTERVAL 0 // With a large buffer, skipping is almost eliminated with 0
 #endif
 #ifdef MAX_PL_READ_BYTES
- //#define MAX_PL_READ_BYTES 65536 // Makes chunked HTML transfers... possibly enable if experiencing slowdowns when sending web UI to client. Untested.
+  //#define MAX_PL_READ_BYTES 1024*64 // Makes chunked HTML transfers but only for playlist.csv and wifi.csv... might help I2S decoder with large playlists
 #endif
 /* Maximum lengths of character buffers */
 /* Notes have been made regarding old yoRadio values */
@@ -823,32 +836,29 @@ https://trip5.github.io/ehRadio/myoptions/generator.html
   #define STREAM_TIMEOUT_MS 3000
 #endif
 
-// PSRAM Buffer
+// PSRAM Audio Buffer
+// WebUI files cache: (up to) 300KB + Decoder Output: 52KB + Display framebuffer (480×320): 74KB = 426KB... leaves 598KB free in 1MB or 1598 free in 2MB
+// To be fair, in real-world use, above 300KB may be excessive...
 #ifndef PSRAM_BUFSIZE
-  #if defined(ARDUINO_ESP32S3_DEV)
-    #define PSRAM_BUFSIZE UINT16_MAX * 10 // ~655KB
+  #if defined(ARDUINO_ESP32_DEV)
+    #define PSRAM_BUFSIZE 550
   #else
-    #define PSRAM_BUFSIZE UINT16_MAX * 20 // 1310KB
+    #define PSRAM_BUFSIZE 1500
   #endif
 #endif
-// FLAC Reserved Buffer
-#ifndef PSRAM_RES_BUFSIZE
-  #if defined(ARDUINO_ESP32S3_DEV)
-    #define PSRAM_RES_BUFSIZE 4096 * 6 // 24KB
-  #else
-    #define PSRAM_RES_BUFSIZE 4096 * 18 // 72KB
-  #endif
+// Since the audio buffer can never technically fill up (most streams prevent reading-ahead excessively),
+// the visual buffer bar should look "full" when the buffer is near to maximum potential (250KB is ~6 seconds at 320kbps)
+#ifndef BUFFERBAR_VISUAL_FULL_KB
+  #define BUFFERBAR_VISUAL_FULL_KB 250
 #endif
-
-// Since the buffer can never technically fill up, the visual buffer bar should look "full"
-// when the buffer is near to maximum potential, estimating that 80% to 85% is good.
-#ifndef BUFFERBAR_VISUAL_FULL_PERCENT // value where buffer bar appears visually full
-  #define BUFFERBAR_VISUAL_FULL_PERCENT 82
+#if (BUFFERBAR_VISUAL_FULL_KB < 16) || (BUFFERBAR_VISUAL_FULL_KB > 500)
+  #warning BUFFERBAR_VISUAL_FULL_KB is outside 16 to 500 range, using default of 250KB
+  #undef BUFFERBAR_VISUAL_FULL_KB
+  #define BUFFERBAR_VISUAL_FULL_KB 250
 #endif
-#if BUFFERBAR_VISUAL_FULL_PERCENT < 1 || BUFFERBAR_VISUAL_FULL_PERCENT > 100
-  #warning BUFFERBAR_VISUAL_FULL_PERCENT out of range, resetting to default 82
-  #undef BUFFERBAR_VISUAL_FULL_PERCENT
-  #define BUFFERBAR_VISUAL_FULL_PERCENT 82
+#if (PSRAM_BUFSIZE < BUFFERBAR_VISUAL_FULL_KB)
+  #undef BUFFERBAR_VISUAL_FULL_KB
+  #define BUFFERBAR_VISUAL_FULL_KB (PSRAM_BUFSIZE)
 #endif
 
 /* --- CPU CORES --- */
@@ -1280,6 +1290,9 @@ https://trip5.github.io/ehRadio/myoptions/generator.html
   #ifndef CORE_MONITOR
     #define CORE_MONITOR // This shows the ESP32 CPU Core Monitor in serial and telnet (updated every 5s), includes SPIFFS information
   #endif
+  #ifndef CORE_MONITOR_ETC_LOOPS
+    #define CORE_MONITOR_ETC_LOOPS 5 // Show SPIFFS + PSRAM info every N core monitor cycles (default: 5, every 25s)
+  #endif
   #ifdef WIDGET_DEBUG
     #define WIDGET_DEBUG // This shows the Widget's Text in logging.  It's extremely noisy.
   #endif
@@ -1288,33 +1301,45 @@ https://trip5.github.io/ehRadio/myoptions/generator.html
  // This enables CORS policy: 'Access-Control-Allow-Origin' (for testing)
 #endif
 
-/* CPU CORE NAMES: Name the Cores For What They Do */
+/* CPU CORE NAMES: Name the Cores for what they do... */
 /* Do not use any of these macros directly in myoptions.h! */
 /* Use #define CORE_MONITOR to show how busy the cores are... this gives them friendly names */
 #if (!CONFIG_FREERTOS_UNICORE && defined(CORE_MONITOR))
   #if (AUDIO_CORE==0)
-    #define CORE_0A "+Audio"
+    #define CORE_0A "Audio"
     #define CORE_1A ""
   #elif (AUDIO_CORE==1)
     #define CORE_0A ""
     #define CORE_1A "+Audio"
   #endif
   #if (NETWORK_CORE==0)
-    #define CORE_0B "+Net"
+    #if (AUDIO_CORE==0)
+      #define CORE_0B "+Net"
+    #else
+      #define CORE_0B "Net"
+    #endif
     #define CORE_1B ""
   #elif (NETWORK_CORE==1)
     #define CORE_0B ""
     #define CORE_1B "+Net"
   #endif
   #if (CONFIG_ASYNC_TCP_RUNNING_CORE==0)
-    #define CORE_0C "+TCP"
+    #if (AUDIO_CORE==0) || (NETWORK_CORE==0)
+      #define CORE_0C "+TCP"
+    #else
+      #define CORE_0C "TCP"
+    #endif
     #define CORE_1C ""
   #elif (CONFIG_ASYNC_TCP_RUNNING_CORE==1)
     #define CORE_0C ""
     #define CORE_1C "+TCP"
   #endif
   #if (DSP_TASK_CORE_ID==0)
-    #define CORE_0D "+Disp"
+    #if (AUDIO_CORE==0) || (NETWORK_CORE==0) || (CONFIG_ASYNC_TCP_RUNNING_CORE==0)
+      #define CORE_0D "+Disp"
+    #else
+      #define CORE_0D "Disp"
+    #endif
     #define CORE_1D ""
   #elif (DSP_TASK_CORE_ID==1)
     #define CORE_0D ""
@@ -1402,12 +1427,6 @@ https://trip5.github.io/ehRadio/myoptions/generator.html
 /* ============================== USER DEFAULTS ============================== */
 /* Sets defaults but still editable in WebUI */
 
-#ifndef VOLUME_SCALE // Default 42 doubles future upstream library's scale of 21; 50-100 gives extremely fine control
-  #define VOLUME_SCALE 42
-#endif
-#if (VOLUME_SCALE < 21) || (VOLUME_SCALE > 255)
-  #error VOLUME_SCALE must be between 21 and 255
-#endif
 #if defined(SOUND_VOLUME) && ((SOUND_VOLUME < 0) || (SOUND_VOLUME > VOLUME_SCALE))
   #warning "define warning in myoptions.h: SOUND_VOLUME is out of range (0-" #VOLUME_SCALE "), reverting to a safe default"
   #undef SOUND_VOLUME