+++ src/displays/displaySSD1305.h
@@ -5,6 +5,8 @@
 #include <Adafruit_GFX.h>
 #include <Adafruit_SSD1305.h>
 
+#define DISPLAY_MODEL_NAME "SSD1305"
+
 typedef GFXcanvas1 Canvas;
 typedef Adafruit_SSD1305 yoDisplay;