+++ src/displays/displaySSD1306.h
@@ -5,6 +5,8 @@
 #include <Adafruit_GFX.h>
 #include <Adafruit_SSD1306.h>
 
+#define DISPLAY_MODEL_NAME "SSD1306"
+
 typedef GFXcanvas1 Canvas;
 typedef Adafruit_SSD1306 yoDisplay;