+++ src/libraries/VS1053_Audio/audioVS1053Ex.h
@@ -4,6 +4,7 @@
  *  Created on: Jul 09.2017
  *  Updated on: Aug 15.2022, Apr 04.2025 (Maleksm)
  *      Author: Wolle (schreibfaul1), easy
+ *  Some adjustments by Trip5 for ehRadio
  */
 #if VS1053_CS!=255
 #ifndef _vs1053_ext
@@ -152,14 +153,14 @@ public:
 
 protected:
 //    size_t m_buffSizePSRAM    = 327675;   // most webstreams limit the advance to 100...300Kbytes
-    size_t m_buffSizePSRAM    = PSRAM_BUFSIZE;
+    size_t m_buffSizePSRAM    = 1024 * PSRAM_BUFSIZE;
     size_t m_buffSizeRAM      = 1600 * AUDIOBUFFER_MULTIPLIER2;
     size_t       m_buffSize         = 0;
     size_t       m_freeSpace        = 0;
     size_t       m_writeSpace       = 0;
     size_t       m_dataLength       = 0;
     size_t       m_resBuffSizeRAM   = 2048;     // reserved buffspace, >= one mp3  frame
-    size_t   m_resBuffSizePSRAM = PSRAM_RES_BUFSIZE;
+    size_t   m_resBuffSizePSRAM     = 1024 * 24; // FLAC frame size 24KB (minimum)
     size_t       m_maxBlockSize     = 1600;
     uint8_t*     m_buffer           = NULL;
     uint8_t*     m_writePtr         = NULL;